add tests for 'networks'
