--- manifest/testdata/with-params-override-override.yaml
@@ -1,10 +1,0 @@
-deriveFrom: "./testdata/with-params-override.yaml"
-
-package:
-  name: "overriddentwice"
-
-initialBlocks:
-  mod1: 175
-  mod2: 175
-params:
-  mod2: "goodbye"
