--- manifest/testdata/with-params-override.yaml
@@ -1,12 +1,0 @@
-deriveFrom: "./testdata/with-params.yaml"
-
-package:
-  name: "overridden"
-  version: v2.3.5
-
-#network: polygon
-initialBlocks:
-  mod1: 150
-  mod2: 150
-params:
-  mod2: "hello"
